Changed type of angle from float to double

+++ src/stp/new_skills/GoToPos.cpp
@@ -25,7 +25,7 @@ Status GoToPos::onUpdate(const StpInfo &info) noexcept {
     Vector2 targetVelocity = robotCommand.vel.stretchToLength(targetVelocityLength);
 
     // Constrain and set angle between -pi and pi
-    float targetAngle = rtt::ai::control::ControlUtils::constrainAngle(robotCommand.angle) - M_PI;
+    double targetAngle = rtt::ai::control::ControlUtils::constrainAngle(robotCommand.angle) - M_PI;
 
     // Set velocity and angle commands
     command.mutable_vel()->set_x(targetVelocity.x);
